requirements.txt
# Django Core
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Database
psycopg2-binary>=2.9,<3.0
dj-database-url>=2.1,<3.0

# Authentication
djangorestframework-simplejwt>=5.3,<6.0

# Background Tasks
celery>=5.3,<6.0
redis>=5.0,<6.0

# External APIs
google-generativeai>=0.4,<1.0
replicate>=0.25,<1.0
requests>=2.31,<3.0
imagekitio>=3.2,<4.0

# Utilities
python-dotenv>=1.0,<2.0
Pillow>=10.0,<11.0

# Production
gunicorn>=21.0,<23.0
whitenoise>=6.6,<7.0
